+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 1.0.0
 
-- Lightweight Dart package to write CSV data to a `StringSink`.
-- Initial commit.
+- Lightweight Dart package to write CSV data to a `StringSink`. Supports and extends RFC 4180.
+
+